# optimize json tokenizer